refactor files into dataset specific folders
