+++ land_grab/common.py
