+++ land_grab/constants.py
@@ -42,7 +42,7 @@ MERGED_DIRECTORY = 'merged/'
 CLEANED_DIRECTORY = 'cleaned/'
 CESSIONS_DIRECTORY = 'cessions/'
 SUMMARY_STATISTICS_DIRECTORY = 'summary_statistics/'
-SOURCE_DIRECTORY = 'source/'
+SOURCE_DIRECTORY = 'parcel_ID_lists/'
 UNIVERSITY_DATA_SOURCE_DIRECTORY = UNIVERSITY_DIRECTORY + SOURCE_DIRECTORY
 STATE_TRUST_DATA_SOURCE_DIRECTORY = STATE_TRUST_DIRECTORY + SOURCE_DIRECTORY
